+++ src/main/kotlin/io/spiralhouse/cycletime/mcp/websocket/ConnectionFactory.kt
@@ -1,19 +1,21 @@
 package io.spiralhouse.cycletime.mcp.websocket
 
 import io.ktor.websocket.DefaultWebSocketSession
+import java.time.Instant
+import java.util.UUID
+import java.util.concurrent.atomic.AtomicReference
 
 /**
- * Factory interface for creating and configuring connections.
+ * Factory for creating and configuring WebSocket connections.
  * 
- * This abstraction follows the Factory pattern to encapsulate the creation
- * logic for connections, allowing for different connection types and
- * initialization strategies.
+ * Creates WebSocket connections with UUID-based identifiers and
+ * proper initialization of tracking metadata.
  * 
  * ## Design Benefits
  * - Centralized connection creation logic
  * - Consistent connection initialization
- * - Support for different connection types
- * - Testability through mock implementations
+ * - UUID v4 for guaranteed uniqueness
+ * - Atomic references for thread-safe activity tracking
  * 
  * ## Future Extensibility
  * Could support:
@@ -22,7 +24,7 @@ import io.ktor.websocket.DefaultWebSocketSession
  * - Custom connection metadata
  * - Connection recycling
  */
-interface ConnectionFactory {
+class ConnectionFactory {
     
     /**
      * Creates a new connection wrapper for a WebSocket session.
@@ -30,14 +32,24 @@ interface ConnectionFactory {
      * @param session the underlying WebSocket session
      * @return a new connection wrapper with unique ID and tracking
      */
-    fun createConnection(session: DefaultWebSocketSession): ActiveWebSocketSession
+    fun createConnection(session: DefaultWebSocketSession): ActiveWebSocketSession {
+        val now = Instant.now()
+        return ActiveWebSocketSession(
+            id = generateConnectionId(),
+            session = session,
+            connectedAt = now,
+            lastActivity = AtomicReference(now)
+        )
+    }
     
     /**
      * Generates a unique identifier for a new connection.
      * 
      * @return a unique connection identifier
      */
-    fun generateConnectionId(): String
+    fun generateConnectionId(): String {
+        return UUID.randomUUID().toString()
+    }
     
     /**
      * Validates if a connection ID is valid.
@@ -45,5 +57,9 @@ interface ConnectionFactory {
      * @param connectionId the ID to validate
      * @return true if the ID is valid, false otherwise
      */
-    fun isValidConnectionId(connectionId: String): Boolean
+    fun isValidConnectionId(connectionId: String): Boolean {
+        // Use regex validation instead of exception control flow to avoid detekt violation
+        val uuidRegex = Regex("^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$")
+        return connectionId.matches(uuidRegex)
+    }
 }
